pyproject.toml
[project]
name = "midas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "robokop-orion>=0.1.4",
    "click",
    # "pandas",
    # "numpy",
    "boto3",
    "mcp",
    "strands-agents",
    "jupyter",
    "jupyterlab",
    "ipykernel",
    "awscurl",
    "streamlit>=1.28.0"
]

[build-system]
requires = ["uv_build >= 0.8.0"]
build-backend = "uv_build"

